requirements.txt
# Chess Tournament Entry Prediction — Python dependencies
# Install with: pip install -r requirements.txt

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0          # faster HTML parser for BeautifulSoup

# Numerical / modeling
numpy>=1.26.0
scipy>=1.12.0        # curve_fit for double-sigmoid fitting

# Excel file parsing (to extract World Open data from CCA workbook)
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.9   # PostgreSQL adapter

# Task scheduling (optional, for recurring scrapes)
celery>=5.3.0
redis>=5.0.0

# API server (Phase 2+)
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0

# Alerting
sendgrid>=6.11.0
twilio>=8.0.0

# Environment variable management
python-dotenv>=1.0.0
